debugging e2e tests
