+++ tests-e2e/cypress/integration/backstage/playbook_edit_spec.js
@@ -6,7 +6,10 @@
 // - [*] indicates an assertion (e.g. * Check the title)
 // ***************************************************************
 
+import merge from 'deepmerge';
+
 import users from '../../fixtures/users.json';
+import {getDefaultConfig} from '../../support/api/system';
 
 describe.only('backstage playbook details', () => {
     before(() => {
@@ -448,6 +451,22 @@ describe.only('backstage playbook details', () => {
                     // # Create a playbook with a user that is later removed from the team
                     cy.legacyApiLogin('sysadmin')
                         .then(() => {
+                            cy.apiGetConfig().then(({config: currentConfig}) => {
+                                const newConfig = {ServiceSettings: {EnableOnboardingFlow: false}};
+
+                                const oldWayOfMakingConfig = merge.all([currentConfig, newConfig]);
+                                cy.log('Old way of making config:');
+                                cy.log(oldWayOfMakingConfig);
+
+                                const altWayOfMakingConfig = merge.all([getDefaultConfig(), currentConfig, newConfig]);
+                                cy.log('Alt way of making config:');
+                                cy.log(altWayOfMakingConfig);
+
+                                const config = merge.all([currentConfig, getDefaultConfig(), newConfig]);
+                                cy.log('Working way of making config:');
+                                cy.log(config);
+                            });
+
                             cy.apiUpdateConfig({
                                 ServiceSettings: {EnableOnboardingFlow: false},
                             });
